Move add user from form logic to extension Add user from form takes the place of the old author import extension Does away with the idea of self.contributors Saves correctly
