+++ website/static/js/registrationEditorExtensions.js
@@ -7,6 +7,7 @@ var FilesWidget = require('js/filesWidget');
 var Fangorn = require('js/fangorn');
 var $osf = require('js/osfHelpers');
 var waterbutler = require('js/waterbutler');
+var ContribAdder = require('js/contribAdder');
 
 var node = window.contextVars.node;
 
@@ -182,57 +183,47 @@ var AuthorImport = function(data, $root) {
     var self = this;
 
     self.question = data;
-    self.contributors = $root.contributors();
 
-    self.setContributorBoxes = function(value) {
-        var boxes = document.querySelectorAll('#contribBoxes input[type="checkbox"]');
-        $.each(boxes, function(i, box) {
-            this.checked = value;
-        });
+    self.makeContributorsRequest = function() {
+        var self = this;
+        var contributorsUrl = window.contextVars.node.urls.api + 'get_contributors/';
+        return $.getJSON(contributorsUrl);
     };
+    self.getContributors = function() {
+        var self = this;
+        return self.makeContributorsRequest()
+            .then(function(data) {
+                return $.map(
+                    data.contributors, function(c) { return c.fullname; }
+                ).join(', ');
+            }).fail(function() {
+                $osf.growl('Could not retrieve contributors.', 'Please refresh the page or ' +
+                           'contact <a href="mailto: support@cos.io">support@cos.io</a> if the ' +
+                           'problem persists.');
+            });
+    };
+
+    self.getContributors().done(function(data) {
+        self.value(data);
+    });
+
     self.preview = function() {
         return self.value();
     };
 
-    self.authorDialog = function() {
-
-        bootbox.dialog({
-            title: 'Choose which contributors to import:',
-            message: function() {
-                ko.renderTemplate('importContributors', self, {}, this, 'replaceNode');
-            },
-            buttons: {
-                cancel: {
-                    label: 'Cancel',
-                    className: 'btn-default',
-                    callback: bootbox.hideAll
-                },
-                select: {
-                    label: 'Import',
-                    className: 'btn-primary',
-                    callback: function() {
-                        var boxes = document.querySelectorAll('#contribBoxes input[type="checkbox"]');
-                        var authors = [];
-                        $.each(boxes, function(i, box) {
-                            if( this.checked ) {
-                                authors.push(this.value);
-                            }
-                        });
-                        if (authors && authors.length) {
-                            var oldValue = self.question.value();
-                            if (!/^\s*$/.test(oldValue || '')) {
-                                self.question.value(oldValue + ', ' + authors.join(', '));
-                            }
-                            else {
-                                self.question.value(authors.join(', '));
-                            }
-                        }
-                    }
-                }
-
-            }
-        });
+    var callback = function(data) {
+        self.value(
+            $.map(data.contributors, function(c){ return c.fullname; })
+                .join(', '));
     };
+    var adder = new ContribAdder(
+        '#addContributors',
+        node.title,
+        node.id,
+        null,
+        null,
+        {async: true, callback: callback}
+    );
 
     $.extend(self, data);
 };
